Move in house app functions to separate files
